+++ cmd/kube-aggregator/pkg/apis/apiregistration/validation/validation.go
@@ -19,10 +19,10 @@ package validation
 import (
 	"fmt"
 
+	"k8s.io/apimachinery/pkg/api/validation/path"
 	utilvalidation "k8s.io/apimachinery/pkg/util/validation"
 	"k8s.io/apimachinery/pkg/util/validation/field"
 	"k8s.io/kubernetes/pkg/api/validation"
-	"k8s.io/kubernetes/pkg/api/validation/path"
 
 	discoveryapi "k8s.io/kubernetes/cmd/kube-aggregator/pkg/apis/apiregistration"
 )
