+++ pkg/apis/autoscaling/validation/BUILD
@@ -14,10 +14,10 @@ go_library(
     tags = ["automanaged"],
     deps = [
         "//pkg/api/validation:go_default_library",
-        "//pkg/api/validation/path:go_default_library",
         "//pkg/apis/autoscaling:go_default_library",
         "//pkg/apis/extensions/v1beta1:go_default_library",
         "//pkg/controller/podautoscaler:go_default_library",
+        "//vendor:k8s.io/apimachinery/pkg/api/validation/path",
         "//vendor:k8s.io/apimachinery/pkg/util/validation/field",
     ],
 )
