+++ pkg/genericapiserver/registry/rest/BUILD
@@ -23,9 +23,9 @@ go_library(
     deps = [
         "//pkg/api:go_default_library",
         "//pkg/api/validation/genericvalidation:go_default_library",
-        "//pkg/api/validation/path:go_default_library",
         "//vendor:k8s.io/apimachinery/pkg/api/errors",
         "//vendor:k8s.io/apimachinery/pkg/api/meta",
+        "//vendor:k8s.io/apimachinery/pkg/api/validation/path",
         "//vendor:k8s.io/apimachinery/pkg/apis/meta/internalversion",
         "//vendor:k8s.io/apimachinery/pkg/apis/meta/v1",
         "//vendor:k8s.io/apimachinery/pkg/runtime",
