+++ pkg/storage/util.go
@@ -25,9 +25,9 @@ import (
 	"github.com/golang/glog"
 
 	"k8s.io/apimachinery/pkg/api/meta"
+	"k8s.io/apimachinery/pkg/api/validation/path"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/util/validation/field"
-	"k8s.io/kubernetes/pkg/api/validation/path"
 )
 
 type SimpleUpdateFunc func(runtime.Object) (runtime.Object, error)
